Don't hard link over existing files
Also, check symlinks at the end.

+++ download-data.sh
@@ -7,8 +7,6 @@ URL=${OPENPBTA_URL:-https://s3.amazonaws.com/kf-openaccess-us-east-1-prd-pbta/da
 RELEASE=${OPENPBTA_RELEASE:-release-v11-20191126}
 PREVIOUS=${OPENPBTA_RELEASE:-release-v10-20191115}
 
-# Remove symlinks in data
-find data -type l -delete
 
 # The md5sum file provides our single point of truth for which files are in a release.
 curl --create-dirs $URL/$RELEASE/md5sum.txt -o data/$RELEASE/md5sum.txt -z data/$RELEASE/md5sum.txt
@@ -28,8 +26,11 @@ then
   # Hard link unchanged files
   for oldfile in "${UNCHANGED[@]}"
   do
-    echo "Hard linking $oldfile"
-    ln data/$PREVIOUS/$oldfile data/$RELEASE/$oldfile
+    if [ ! -e "data/$RELEASE/$oldfile" ]
+    then
+      echo "Hard linking $oldfile"
+      ln data/$PREVIOUS/$oldfile data/$RELEASE/$oldfile
+    fi
   done
 fi
 
@@ -71,6 +72,9 @@ cd -
 cd data/$RELEASE
 md5sum -c md5sum.txt
 cd ../../
+
+# Remove old symlinks in data
+find data -type l -delete
 
 # Make symlinks in data/ to the files in the just downloaded release folder.
 for file in "${FILES[@]}"
